+++ src/vscode/treeviews/workspaces/DatabricksWorkspaceFile.ts
@@ -185,7 +185,7 @@ export class DatabricksWorkspaceFile extends DatabricksWorkspaceTreeItem {
 	}
 
 	async compare(): Promise<void> {
-		let onlineFileTempPath: vscode.Uri = vscode.Uri.parse("dbws:" + this.path);
+		let onlineFileTempPath: vscode.Uri = vscode.Uri.parse(ThisExtension.WORKSPACE_SCHEME + ":" + this.path);
 
 		Helper.showDiff(onlineFileTempPath, this.localPath);
 	}